Skip rendering trailing whitespace within each row
Whitespace at the end of a row does not contribute to its visual
appearance since we will end it with a line break anyway. For some
applications the additional whitespace is problematic since it
unnecessarily bloats the size of the rendered output.

This commit extends carrot to skip all trailing whitespace glyphs
in each row to avoid this problem. A whitespace glyph is currently
defined as a space character with the default style of the target.