Explicitly pass the used locale as part of the target
Instead of using a fixed (global) locale, different locales
can now be used within the same application. The ability to
set the locale for each target will be added in a follow-up commit.

Since the locale might influence the layout of the blocks,
each form does now depend on a specific target.

To facilitate the propagation of the target during the rendering
process, the block interface has been changed slightly. Notably,
the extent of a block does now depend on the target.

The commit also simplifies the rendering test case via
Catch's sections.