Switch to a Conan-based CI setup for Linux
We already use Conan to build the CI dependencies
for Windows which allows us to easily test different
library configurations without building a new Docker image
for every possible combination.

All jobs are now based on the latest Ubuntu image which
allows us to use a suitable set of compilers for the tests.

[skip appveyor]