Document the error handling for each function
In particular, every external function is now
either marked as noexcept or the exceptions it
can throw are documented.

Due to the fact, that we only throw exceptions
on recoverable errors, this means that most
functions are now noexcept. For a few of the
remaining uses of exceptions, it it still debatable
if one can recover from these errors.