Modernize the build process
To simplify the build process, we now assume
that all dependencies will be installed via
Conan. Since Conan currently does not generate
the same targets as the CMake-generated config
files, this unfortunately means that building
carrot using dependencies installed via other
means, including system packages, will be
non-trivial. We hope as the support for
components in Conan matures, that we can
restore this ability. Since carrot is currently
primarily deployed via Conan, I consider this
a sensible trade-off.

In addition, this commit updates all dependencies
to use Conan Center as the source.